backend/requirements.txt
# Django核心
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.0

# 环境变量
python-dotenv>=1.0

# 数据库
pymysql>=1.0
sqlalchemy>=2.0

# WebSocket
channels>=4.0
channels-redis>=4.0

# Celery异步任务
celery>=5.3
redis>=5.0

# 缓存
django-redis>=5.4

# 数据处理
pandas>=2.0
numpy>=1.26
openpyxl>=3.0
bokeh>=3.0
matplotlib>=3.8

# 安全
cryptography>=40.0

# 图片处理
Pillow>=10.0

# HTTP
requests>=2.31

# ASGI服务器
daphne>=4.0

# Celery Beat调度器
django-celery-beat>=2.5
